feat: add Docker support and refactor project structure
- Added Dockerfile for containerized deployment
 - Multi-stage build for optimized image size
 - Poetry for dependency management
 - Entrypoint script for running tests and starting the API
- Updated CI/CD workflows
 - Added `--no-root` flag to Poetry install commands in GitHub and GitLab CI
 - Upgraded Poetry version to 2.1.1 in GitLab CI
- Refactored project structure
 - Moved FastAPI app logic to `src/app.py`
 - Extracted segment tree logic to `src/segment_tree.py`
 - Created separate schemas module in `src/schemas.py`
 - Added conftest.py for test fixtures
- Updated .gitignore
 - Added `.idea/` directory to ignore list
- Updated README.md
 - Added Docker build and run instructions
 - Added test commands for Docker
- Split tests into separate modules
 - Moved schema tests to `src/tests/test_schemas.py`
 - Moved segment tree tests to `src/tests/test_segment_tree.py`
 - Simplified `src/tests/test_app.py` to focus on API tests
