+++ .github/workflows/test.yml
@@ -37,7 +37,7 @@ jobs:
 
     - name: Install dependencies
       if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-      run: poetry install --no-interaction --with dev
+      run: poetry install --no-interaction --with dev --no-root
 
     - name: Run tests
       run: |
